handle validate form worksheet

+++ CMS/src/form/FormWorksheet/FormWorksheetAddCashier.js
@@ -9,6 +9,8 @@ import {
 	updateWorksheetByIDAsync,
 } from '../../redux/worksheet/action';
 import {useDispatch, useSelector} from 'react-redux';
+import {confirmModal, errorAlert} from '../../components/Notify/Alert';
+import {succesNotify} from '../../components/Notify/Toast';
 
 const FormWorksheetAddCashier = ({handleGetWorkSheet, handleAddClick, handleToggleForm, add}) => {
 	FormWorksheetAddCashier.propTypes = {
@@ -54,6 +56,11 @@ const FormWorksheetAddCashier = ({handleGetWorkSheet, handleAddClick, handleTogg
 		}
 	}, [worksheetDetailsFromRedux]);
 
+	//SHOW TOAST
+	const showSuccessToast = (text) => {
+		succesNotify(text);
+	};
+
 	//HADNLE ADD WORKSHEET
 	const handleSubmit = () => {
 		// dispatch(addTaskGuard(name, worksheet, sheet));
@@ -79,11 +86,15 @@ const FormWorksheetAddCashier = ({handleGetWorkSheet, handleAddClick, handleTogg
 				setWorksheet('');
 				setSheet('');
 				setError('');
+				//Show toast
+				showSuccessToast('Asign employee successfully!');
 				// REFRESH WORKSHEET
 				handleToggleForm();
 			})
 			.catch((error) => {
 				console.log(error);
+				let errorMessage = error.response.data.message.split(':')[1].trim();
+				errorAlert(errorMessage.split('Employee_name')[0]);
 			});
 	};
 
@@ -104,31 +115,48 @@ const FormWorksheetAddCashier = ({handleGetWorkSheet, handleAddClick, handleTogg
 			})
 			.then((response) => {
 				dispatch(fetchWorksheetByID(0));
+				//Show toast
+				showSuccessToast('Update successfully!');
 				handleToggleForm();
 			})
 			.catch((error) => {
 				console.log(error);
+				let errorMessage = error.response.data.message.split('-')[1].trim();
+				errorAlert(errorMessage.split('Employee_name')[0]);
 			});
 	};
 
 	// HANDLE DELETE WORKSHEET
-	const handleDelete = () => {
+	const handleDelete = async () => {
 		console.log('vao handleDelete');
-		const idToDelete = worksheetDetail.id;
-		dispatch(fetchWorksheetByID(0));
-		dispatch(deleteWorksheetByIDAsync(idToDelete)).then((response) => {
-			const startDate = calenderDay[0].date;
-			const endDate = calenderDay[calenderDay.length - 1].date;
-			handleGetWorkSheet(`${startDate},${endDate}`);
-		});
-		handleToggleForm();
+		let isDelete = await confirmModal('Yes, delete it')
+			.then((isConfirmed) => {
+				return isConfirmed;
+				// console.log('Confirmation result:', isConfirmed); // Output: Confirmation result: true
+			})
+			.catch((error) => {
+				console.error('Confirmation error:', error);
+			});
+		try {
+			if (isDelete) {
+				const idToDelete = worksheetDetail.id;
+				dispatch(fetchWorksheetByID(0));
+				dispatch(deleteWorksheetByIDAsync(idToDelete)).then((response) => {
+					const startDate = calenderDay[0].date;
+					const endDate = calenderDay[calenderDay.length - 1].date;
+					handleGetWorkSheet(`${startDate},${endDate}`);
+				});
+				handleToggleForm();
+			}
+		} catch (error) {
+			console.log(error);
+		}
 	};
 
 	if (!listName) {
 		return;
 	}
 
-	console.log('worksheetDetail', worksheetDetail);
 	return (
 		<div className={styles.formWorksheet}>
 			<div className={styles.formContainer}>
@@ -154,7 +182,7 @@ const FormWorksheetAddCashier = ({handleGetWorkSheet, handleAddClick, handleTogg
 							{/* Add more options as needed */}
 						</select>
 					</div>
-					{worksheetDetail ? (
+					{!add ? (
 						<div></div>
 					) : (
 						<div>
